requirements.txt
pygame
pytiled-parser
pillow